refactor: move tests to their command files for users
